Common service collection extensions.
