+++ src/Lib/Kros.ToDos.Base/Kros.ToDos.Base.csproj
@@ -6,8 +6,12 @@
 
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="FluentValidation.AspNetCore" Version="10.3.0" />
     <PackageReference Include="Kros.AspNetCore" Version="2.41.0" />
+    <PackageReference Include="Kros.Swagger.Extensions" Version="1.5.0" />
+    <PackageReference Include="MicroElements.Swashbuckle.FluentValidation" Version="5.2.0" />
     <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="5.0.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.1.5" />
   </ItemGroup>
 
   <ItemGroup>
