+++ src/Services/Kros.Authorization.Api/Application/ServiceCollectionExtensions.cs
@@ -1,14 +1,7 @@
-﻿using FluentValidation.AspNetCore;
-using Kros.AspNetCore.Authorization;
+﻿using Kros.AspNetCore.Authorization;
 using Kros.KORM.Extensions.Asp;
-using Kros.Swagger.Extensions;
-using Kros.ToDos.Base.Infrastructure;
 using MediatR;
-using MicroElements.Swashbuckle.FluentValidation.AspNetCore;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.IO;
 using System.Reflection;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -18,18 +11,6 @@ namespace Microsoft.Extensions.DependencyInjection
     /// </summary>
     public static class ServiceCollectionExtensions
     {
-        /// <summary>
-        /// Register fluent validation.
-        /// </summary>
-        /// <param name="builder">MVC builder.</param>
-        /// <returns>MVC builder.</returns>
-        public static IMvcCoreBuilder AddFluentValidation(this IMvcCoreBuilder builder)
-            => builder.AddFluentValidation(o =>
-            {
-                o.RegisterValidatorsFromAssembly(Assembly.GetExecutingAssembly());
-                o.DisableDataAnnotationsValidation = true;
-            });
-
         /// <summary>
         /// Add KORM database.
         /// </summary>
@@ -52,25 +33,6 @@ namespace Microsoft.Extensions.DependencyInjection
             => services.AddMediatR(Assembly.GetExecutingAssembly());
 
         /// <summary>
-        /// Add Swagger.
-        /// </summary>
-        /// <param name="services">DI container.</param>
-        /// <param name="configuration">Application configuration.</param>
-        public static IServiceCollection AddSwagger(this IServiceCollection services, IConfiguration configuration)
-            => services
-                .AddSwaggerDocumentation(configuration, c =>
-                {
-                    c.EnableAnnotations();
-
-                    string xmlDocFilePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Kros.Authorization.Api.xml");
-                    if (File.Exists(xmlDocFilePath))
-                    {
-                        c.IncludeXmlComments(xmlDocFilePath);
-                    }
-                })
-                .AddFluentValidationRulesToSwagger();
-
-        /// <summary>
         /// Add application services and configure options.
         /// </summary>
         /// <param name="services">DI container.</param>
@@ -85,50 +47,6 @@ namespace Microsoft.Extensions.DependencyInjection
                 scan.FromCallingAssembly()
                     .AddClasses()
                     .AsMatchingInterface());
-        }
-
-        /// <summary>
-        /// Configure api authorization.
-        /// </summary>
-        /// <param name="services">Collection of app services.</param>
-        /// <param name="scheme">Scheme name for authentication.</param>
-        /// <returns></returns>
-        public static IServiceCollection AddApiJwtAuthorization(
-            this IServiceCollection services,
-            string scheme)
-        {
-            return services.AddAuthorization(options =>
-            {
-                options.AddPolicy(PoliciesHelper.OwnerAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.OwnerRole);
-                });
-
-                options.AddPolicy(PoliciesHelper.AdminAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.AdminRole,
-                                                                                PermissionsHelper.ClaimValues.OwnerRole);
-                });
-
-                options.AddPolicy(PoliciesHelper.WriterAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.AdminRole,
-                                                                                PermissionsHelper.ClaimValues.OwnerRole,
-                                                                                PermissionsHelper.ClaimValues.WriterRole);
-                });
-
-                options.AddPolicy(PoliciesHelper.ReaderAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.AdminRole,
-                                                                                PermissionsHelper.ClaimValues.OwnerRole,
-                                                                                PermissionsHelper.ClaimValues.WriterRole,
-                                                                                PermissionsHelper.ClaimValues.ReaderRole);
-                });
-            });
         }
     }
 }
