+++ src/Services/Kros.ToDos.Api/Application/ServiceCollectionExtensions.cs
@@ -1,18 +1,12 @@
-﻿using FluentValidation.AspNetCore;
-using Kros.KORM.Extensions.Asp;
+﻿using Kros.KORM.Extensions.Asp;
 using Kros.MediatR.Extensions;
-using Kros.Swagger.Extensions;
 using Kros.ToDos.Api.Application;
 using Kros.ToDos.Api.Application.Commands.PipeLines;
 using Kros.ToDos.Api.Application.Queries.PipeLines;
 using Kros.ToDos.Api.Infrastructure;
-using Kros.ToDos.Base.Infrastructure;
 using MediatR;
-using MicroElements.Swashbuckle.FluentValidation.AspNetCore;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Configuration;
-using System;
-using System.IO;
 using System.Reflection;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -22,18 +16,6 @@ namespace Microsoft.Extensions.DependencyInjection
     /// </summary>
     public static class ServiceCollectionExtensions
     {
-        /// <summary>
-        /// Register fluent validation.
-        /// </summary>
-        /// <param name="builder">MVC builder.</param>
-        /// <returns>MVC builder.</returns>
-        public static IMvcCoreBuilder AddFluentValidation(this IMvcCoreBuilder builder)
-            => builder.AddFluentValidation(o =>
-            {
-                o.RegisterValidatorsFromAssembly(Assembly.GetExecutingAssembly());
-                o.DisableDataAnnotationsValidation = true;
-            });
-
         /// <summary>
         /// Add KORM database.
         /// </summary>
@@ -59,25 +41,6 @@ namespace Microsoft.Extensions.DependencyInjection
                 .AddMediatRNullCheckPostProcessor();
 
         /// <summary>
-        /// Add Swagger.
-        /// </summary>
-        /// <param name="services">DI container.</param>
-        /// <param name="configuration">Application configuration.</param>
-        public static IServiceCollection AddSwagger(this IServiceCollection services, IConfiguration configuration)
-            => services
-                .AddSwaggerDocumentation(configuration, c =>
-                {
-                    c.EnableAnnotations();
-
-                    string xmlDocFilePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Kros.ToDos.Api.xml");
-                    if (File.Exists(xmlDocFilePath))
-                    {
-                        c.IncludeXmlComments(xmlDocFilePath);
-                    }
-                })
-                .AddFluentValidationRulesToSwagger();
-
-        /// <summary>
         /// Add distributed cache.
         /// </summary>
         /// <param name="services">DI container.</param>
@@ -101,50 +64,6 @@ namespace Microsoft.Extensions.DependencyInjection
             }
 
             return services;
-        }
-
-        /// <summary>
-        /// Configure api authorization.
-        /// </summary>
-        /// <param name="services">Collection of app services.</param>
-        /// <param name="scheme">Scheme name for authentication.</param>
-        /// <returns></returns>
-        public static IServiceCollection AddApiJwtAuthorization(
-            this IServiceCollection services,
-            string scheme)
-        {
-            return services.AddAuthorization(options =>
-            {
-                options.AddPolicy(PoliciesHelper.OwnerAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.OwnerRole);
-                });
-
-                options.AddPolicy(PoliciesHelper.AdminAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.OwnerRole,
-                                                                                PermissionsHelper.ClaimValues.AdminRole);
-                });
-
-                options.AddPolicy(PoliciesHelper.WriterAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.OwnerRole,
-                                                                                PermissionsHelper.ClaimValues.AdminRole,
-                                                                                PermissionsHelper.ClaimValues.WriterRole);
-                });
-
-                options.AddPolicy(PoliciesHelper.ReaderAuthPolicyName, policyAdmin =>
-                {
-                    policyAdmin.AuthenticationSchemes.Add(scheme);
-                    policyAdmin.RequireClaim(PermissionsHelper.Claims.UserRole, PermissionsHelper.ClaimValues.OwnerRole,
-                                                                                PermissionsHelper.ClaimValues.AdminRole,
-                                                                                PermissionsHelper.ClaimValues.WriterRole,
-                                                                                PermissionsHelper.ClaimValues.ReaderRole);
-                });
-            });
         }
     }
 }
